+++ projects/ng2-ya-table/src/lib/directives/ng2-ya-table-cell-template.directive.ts
