+++ projects/ng2-ya-table/src/lib/directives/ng2-ya-table-sorting.directive.spec.ts
