+++ projects/ng2-ya-table/src/lib/filtering/ng2-ya-table-filtering-list.component.html
