+++ projects/ng2-ya-table/src/lib/ng2-ya-table.module.ts
@@ -3,11 +3,11 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { PaginationModule } from 'ngx-bootstrap/pagination';
 
-import { Ng2YaTableFilteringDefaultComponent } from './ng2-ya-table-filtering-default.component';
-import { Ng2YaTableFilteringListComponent } from './ng2-ya-table-filtering-list.component';
-import { Ng2YaTableSortingDirective } from './ng2-ya-table-sorting.directive';
+import { Ng2YaTableFilteringDefaultComponent } from './filtering/ng2-ya-table-filtering-default.component';
+import { Ng2YaTableFilteringListComponent } from './filtering/ng2-ya-table-filtering-list.component';
+import { Ng2YaTableSortingDirective } from './directives/ng2-ya-table-sorting.directive';
+import { Ng2YaTableCellTemplateDirective } from './directives/ng2-ya-table-cell-template.directive';
 import { Ng2YaTableComponent } from './ng2-ya-table.component';
-import { Ng2YaTableCellTemplateDirective } from './ng2-ya-table-cell-template.directive';
 
 @NgModule({
   declarations: [
