+++ projects/ng2-ya-table/src/public-api.ts
@@ -2,8 +2,8 @@
  * Public API Surface of ng2-ya-table
  */
 
-export * from './lib/ng2-ya-table.service';
-export * from './lib/ng2-ya-table-cell-template.directive';
+export * from './lib/services/ng2-ya-table.service';
+export * from './lib/directives/ng2-ya-table-cell-template.directive';
 export * from './lib/ng2-ya-table.component';
 export * from './lib/ng2-ya-table-interfaces';
 export * from './lib/ng2-ya-table.module';
